requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
uuid_utils
faiss-cpu
python-dotenv
pypdf
camelot-py
opencv-python-headless
faster-whisper
regex
sentence-transformers